+++ torch/csrc/jit/passes/onnx/fold_if_node.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <torch/csrc/jit/ir/ir.h>
+
+namespace torch {
+namespace jit {
+
+void FoldIfONNX(Block* b, bool dynamic_axes);
+
+} // namespace jit
+
+} // namespace torch
