+++ torch/onnx/utils.py
@@ -205,7 +205,9 @@ def _optimize_graph(graph, operator_export_type, _disable_torch_constant_prop=Fa
 
         torch._C._jit_pass_onnx_scalar_type_analysis(graph)
         torch._C._jit_pass_lint(graph)
-
+        
+        dynamic_fold = (dynamic_axes != {})
+        torch._C._jit_pass_onnx_fold_if(graph, dynamic_fold)
         from torch.onnx.symbolic_helper import _export_onnx_opset_version
         torch._C._jit_pass_onnx_peephole(graph, _export_onnx_opset_version, fixed_batch_size)
         torch._C._jit_pass_lint(graph)
